src/requirements.txt
Django==4.0.6
django-environ==0.9.0
mysqlclient==2.1.1
openpyxl==3.0.10
djangorestframework~=3.13.1